ebpf plugin + parser + component test#1 - moving  bpf_event_writer
